Python/Streamlit/OCR app/requirements.txt
streamlit==1.12.0
paddleocr==2.4.0.2
Pillow==9.0.1
opencv-python==4.5.5.64